move gias helpers into GIAS::School model
