+++ app/models/gias/school.rb
@@ -1,5 +1,5 @@
 class GIAS::School < ApplicationRecord
-  include GIAS::Helpers
+  include GIAS::Types
 
   self.table_name = "gias_schools"
 
@@ -40,6 +40,9 @@ class GIAS::School < ApplicationRecord
               message: "must be between 0 and 700,000",
             }
 
+  validates :establishment_number,
+            numericality: { only_integer: true }
+
   validates :local_authority_code,
             numericality: { only_integer: true }
 
@@ -53,9 +56,6 @@ class GIAS::School < ApplicationRecord
               less_than_or_equal_to: 1_300_000,
               message: "must be between 0 and 1,300,000"
             }
-
-  validates :number,
-            numericality: { only_integer: true }
 
   validates :phase_code,
             numericality: {
@@ -88,5 +88,23 @@ class GIAS::School < ApplicationRecord
             },
             uniqueness: true
 
+  # Scopes
+  scope :cip_only, -> { open.where(type_code: GIAS::Types::CIP_ONLY_TYPE_CODES) }
+  scope :eligible_for_funding, -> { open.in_england.where(type_code: GIAS::Types::ELIGIBLE_TYPE_CODES).or(open.in_england.section_41) }
+  scope :in_england, -> { where("administrative_district_code ILIKE 'E%' OR administrative_district_code = '9999'") }
+  scope :open, -> { open_status.or(proposed_to_close_status) }
+  scope :section_41, -> { where(section_41_approved: true) }
+
   # Instance Methods
+  def closed?
+    !open?
+  end
+
+  def in_england?
+    english_district?(administrative_district_code)
+  end
+
+  def open?
+    open_status? || proposed_to_close_status?
+  end
 end
