+++ spec/models/school_spec.rb
@@ -5,8 +5,9 @@ describe School do
   end
 
   describe "validations" do
-    subject { FactoryBot.build(:teacher) }
+    subject { FactoryBot.build(:school) }
 
-    it { is_expected.to validate_presence_of(:name) }
+    it { is_expected.to validate_presence_of(:urn) }
+    it { is_expected.to validate_uniqueness_of(:urn) }
   end
 end
